This commit was generated by cvs2svn to compensate for changes in r5821, which included commits to RCS files with non-trunk default branches.
git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk@5822 379a1393-f5fb-40a0-bcee-ef074d9b53f7

+++ vice/src/petcat.c
@@ -112,6 +112,7 @@
 
 #include "charsets.h"		/* ctrl1, ctrl2, cbmkeys */
 #include "p00.h"
+#include "types.h"
 
 #ifndef SEEK_SET
 #define SEEK_SET 0
@@ -656,7 +657,7 @@ int main(int argc, char **argv)
 
 	    if (p00_check_name(outfilename) >= 0) {
 		printf("writing PC64 header.\n");
-		p00_write_header(dest, (BYTE *)argv[0], 0);
+		p00_write_header(dest, (BYTE *)argv[0], (BYTE)0);
 	    }
 
 	    p_tokenize(version, load_addr, ctrls);
@@ -1071,7 +1072,7 @@ static int   p_expand(version, addr, ctrls)
 		continue;
 	    }
 
-	    _p_toascii(c, ctrls);	/* convert character */
+	    _p_toascii((int)c, ctrls);	/* convert character */
 
 	} while ((c = getc(source)) != EOF && c);
 	printf("\n");
